refactor(frontend): extract JobCard + Pill + status maps (8/N)
The big one. JobCard was the largest single component in App.jsx (300+
LOC of dense JSX with 16 props and 5 shared helpers). Bundled here:

- components/JobCard.jsx  (~310 LOC, memoized export at the bottom)
- components/Pill.jsx     (~10 LOC pure presentational)

Plus extends lib/jobConstants.js with the remaining shared identities
JobCard needs:
- ST_COLOR / ST_LABEL    — application-status colors + labels
- JC_STYLES              — frozen styles object (memo-stable identity)
- isPlatinum, isHighComp — tier helpers used by Pipeline + Rare tabs too
- likelySponsor          — H1B heuristic (was in App.jsx; preserved exactly)

App.jsx changes:
- Removed JC_STYLES block (~25 LOC)
- Removed JobCard + _JobCard + memo wrap (~310 LOC)
- Removed Pill (~3 LOC)
- Removed ST_COLOR + ST_LABEL (~2 LOC)
- Removed isPlatinum, isHighComp, likelySponsor (~10 LOC)
- Added 3 import lines
- Net: 3452 → 3082 LOC (-370)

Build clean, identical bundle size.

The remaining 3082 LOC of App.jsx is the orchestrator (state hooks,
useJobData, useApplications) + the 9 tab render blocks. Each of those
is the next batch of PRs.
